Refactor interface and impl NameScope importing to resemble classes (#5585)
Similar to the class change made in
c6f25e90185f497cfd1cddd5ef87be4aa19238b5 but without tests as it doesn't
appear that the same bug is reachable for these cases at the moment -
but seems good to match the approach in case cycles appear in the future
and to make the code consistent.

The impl case didn't seem to be able to use the new common utility
function, since it splits the two pieces of work and only does the
second conditionally.