Fix the hardening mode defines, also use debug (#5666)
This is based on #5664 because it's fixing an issue which `DEBUG` would
catch. That's also why I'm switching to `DEBUG` from `EXTENSIVE`; I
think we should be okay with the performance cost in `file_test`, which
is probably our main concern.

Note digging into this also got me to notice that the flags weren't
actually enabled; this is fixing the define names.

---------

Co-authored-by: Dana Jansens <danakj@orodu.net>